+++ tests/TestCase/Command/LinkScannerCommandTest.php
@@ -42,12 +42,15 @@ class LinkScannerCommandTest extends TestCase
     protected $fullBaseUrl = 'http://google.com';
 
     /**
-     * Internal method to set and get the `LinkScannerCommand` instance and all
-     *  properties of this test class
+     * Called before every test method
+     * @return void
      */
-    protected function setLinkScannerCommand()
+    public function setUp()
     {
-        $this->LinkScanner = new LinkScanner($this->fullBaseUrl, $this->getClientReturnsSampleResponse());
+        parent::setUp();
+
+        $this->LinkScanner = new LinkScanner($this->getClientReturnsSampleResponse());
+        $this->LinkScanner->setConfig('fullBaseUrl', $this->fullBaseUrl);
         $this->getEventManager();
 
         $this->_out = new ConsoleOutput;
@@ -60,17 +63,6 @@ class LinkScannerCommandTest extends TestCase
     }
 
     /**
-     * Called before every test method
-     * @return void
-     */
-    public function setUp()
-    {
-        $this->setLinkScannerCommand();
-
-        parent::setUp();
-    }
-
-    /**
      * Test for `scan()` method
      * @test
      */
@@ -80,9 +72,10 @@ class LinkScannerCommandTest extends TestCase
 
         $this->assertEquals([
             'cache' => true,
-            'excludeLinks' => ['\{.*\}', 'javascript:'],
+            'excludeLinks' => '/[\{\}+]/',
             'externalLinks' => true,
             'followRedirects' => false,
+            'fullBaseUrl' => $this->fullBaseUrl,
             'maxDepth' => 1,
             'lockFile' => true,
             'target' => TMP . 'cakephp-link-scanner',
@@ -113,7 +106,8 @@ class LinkScannerCommandTest extends TestCase
         }
 
         //With an error response (404 status code)
-        $this->Command->LinkScanner = new LinkScanner($this->fullBaseUrl, $this->getClientReturnsErrorResponse());
+        $this->Command->LinkScanner = new LinkScanner($this->getClientReturnsErrorResponse());
+        $this->Command->LinkScanner->setConfig('fullBaseUrl', $this->fullBaseUrl);
         $this->Command->run(['--verbose'], $this->io);
         $this->assertErrorContains('404');
     }
@@ -146,9 +140,10 @@ class LinkScannerCommandTest extends TestCase
 
         $expectedConfig = [
             'cache' => true,
-            'excludeLinks' => ['\{.*\}', 'javascript:'],
+            'excludeLinks' => '/[\{\}+]/',
             'externalLinks' => true,
             'followRedirects' => false,
+            'fullBaseUrl' => $this->fullBaseUrl,
             'maxDepth' => 2,
             'lockFile' => false,
             'target' => TMP . 'cakephp-link-scanner',
@@ -180,20 +175,22 @@ class LinkScannerCommandTest extends TestCase
         //With a different full base url
         array_pop($params);
         $params[] = '--full-base-url=http://anotherFullBaseUrl';
-        $this->setLinkScannerCommand();
+        self::setUp();
         $this->Command->run($params, $this->io);
+        $expectedConfig['fullBaseUrl'] = 'http://anotherFullBaseUrl';
         $this->assertEquals($expectedConfig, $this->LinkScanner->getConfig());
-        $this->assertEquals('http://anotherFullBaseUrl', $this->LinkScanner->fullBaseUrl);
-        $this->assertOutputRegExp(sprintf('/Scan started for %s/', preg_quote($this->LinkScanner->fullBaseUrl, '/')));
+        $this->assertOutputRegExp(sprintf('/Scan started for %s/', preg_quote($this->LinkScanner->getConfig('fullBaseUrl'), '/')));
         $this->assertErrorEmpty();
 
         //Disables external links
         array_pop($params);
         $params = array_merge($params, ['--full-base-url=' . $this->fullBaseUrl, '--no-external-links']);
 
-        $this->setLinkScannerCommand();
+        self::setUp();
         $this->Command->run($params, $this->io);
-        $this->assertEquals(['externalLinks' => false] + $expectedConfig, $this->LinkScanner->getConfig());
+        $expectedConfig['externalLinks'] = false;
+        $expectedConfig['fullBaseUrl'] = $this->fullBaseUrl;
+        $this->assertEquals($expectedConfig, $this->LinkScanner->getConfig());
 
         $lineDifferentFullBaseUrl = function ($line) {
             $pattern = sprintf('/^Checking https?:\/\/%s/', preg_quote(get_hostname_from_url($this->fullBaseUrl)));
@@ -205,7 +202,7 @@ class LinkScannerCommandTest extends TestCase
 
         //Re-enables external links
         array_pop($params);
-        $this->setLinkScannerCommand();
+        self::setUp();
         $this->Command->run($params, $this->io);
         $expectedConfig['externalLinks'] = true;
         $this->assertEquals($expectedConfig, $this->LinkScanner->getConfig());
@@ -215,7 +212,7 @@ class LinkScannerCommandTest extends TestCase
             'example' => $this->LinkScanner->getConfig('target') . DS . 'example',
             TMP . 'example' => TMP . 'example',
         ] as $filename => $expectedExportFile) {
-            $this->setLinkScannerCommand();
+            self::setUp();
             array_shift($params);
             array_unshift($params, '--export-with-filename=' . $filename);