+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 1.x branch
+### 1.0.4-RC2
+* now it can load a configuration file;
+* removed `LinkScanner::setFullBaseUrl()` method. The full base url can be set
+    as class configuration.
+
 ### 1.0.3-RC1
 * some little fixes.